chore: prettier and rename folder
