+++ axelar-chains-config/index.js
