+++ axelar-chains-config/info/mainnet.json
