+++ axelar-chains-config/info/testnet.json
