+++ axelar-chains-config/tests/getChainArray.test.js
