+++ axelar-chains-config/tests/info.test.js
