+++ axelar-chains-config/tests/schema/index.js
