+++ axelar-chains-config/utils/getChainArray.js
