+++ axelar-chains-config/utils/index.js
@@ -1,2 +1,2 @@
 export * from './getChainArray';
-export * from './readJSON'
+export * from './readJSON';
