+++ axelar-chains-config/utils/readJSON.js
